+++ src/Contracts/X509GeneratorInterface.php
@@ -2,45 +2,49 @@
 
 namespace EbicsApi\Ebics\Contracts;
 
-use EbicsApi\Ebics\Contracts\Crypt\RSAInterface;
 use EbicsApi\Ebics\Contracts\Crypt\X509Interface;
+use EbicsApi\Ebics\Models\X509\X509Context;
 
 /**
  * X509 Factory Interface representation.
  *
  * @license http://www.opensource.org/licenses/mit-license.html  MIT License
- * @author Guillaume Sainthillier
+ * @author Andrew Svirin, Guillaume Sainthillier
  */
 interface X509GeneratorInterface
 {
+    /**
+     * Generate a X509 (Authorization) and returns its content.
+     *
+     * @return X509Interface
+     */
+    public function generateAX509(): X509Interface;
 
     /**
      * Generate a X509 (Authorization) and returns its content.
      *
-     * @param RSAInterface $privateKey
-     * @param RSAInterface $publicKey
-     *
      * @return X509Interface
      */
-    public function generateAX509(RSAInterface $privateKey, RSAInterface $publicKey): X509Interface;
+    public function generateEX509(): X509Interface;
 
     /**
      * Generate a X509 (Authorization) and returns its content.
      *
-     * @param RSAInterface $privateKey
-     * @param RSAInterface $publicKey
-     *
      * @return X509Interface
      */
-    public function generateEX509(RSAInterface $privateKey, RSAInterface $publicKey): X509Interface;
+    public function generateXX509(): X509Interface;
 
     /**
-     * Generate a X509 (Authorization) and returns its content.
-     *
-     * @param RSAInterface $privateKey
-     * @param RSAInterface $publicKey
-     *
+     * Generate issuer X509 using settled public and private key.
      * @return X509Interface
      */
-    public function generateXX509(RSAInterface $privateKey, RSAInterface $publicKey): X509Interface;
+    public function generateIssuerX509(): X509Interface;
+
+    public function getAX509Context(): X509Context;
+
+    public function getEX509Context(): X509Context;
+
+    public function getXX509Context(): X509Context;
+
+    public function getIssuerX509Context(): X509Context;
 }