+++ src/Handlers/OrderDataHandlerV2.php
@@ -8,7 +8,6 @@ use DOMElement;
 use DOMNode;
 use DOMNodeList;
 use EbicsApi\Ebics\Contracts\SignatureInterface;
-use EbicsApi\Ebics\Handlers\Traits\H00XTrait;
 use EbicsApi\Ebics\Models\Crypt\X509;
 use EbicsApi\Ebics\Models\XmlData;
 use EbicsApi\Ebics\Models\XmlDocument;